expose entrypoints into single .so file (#33)
* expose entrypoints into single .so file

* test

* fix

* fix submodule

* fix submodule path

* remove bad file

* whitespace
