+++ Cargo.toml
@@ -104,6 +104,7 @@ members = [
     "svm",
     "svm-callback",
     "svm-feature-set",
+    "svm-fuzz-harness",
     "svm-fuzz-harness/fixture",
     "svm-fuzz-harness/instr",
     "svm-log-collector",
@@ -533,6 +534,7 @@ solana-svm-callback = { path = "svm-callback", version = "=3.1.0" }
 solana-svm-feature-set = { path = "svm-feature-set", version = "=3.1.0" }
 solana-svm-fuzz-harness-fixture = { path = "svm-fuzz-harness/fixture", version = "=3.1.0" }
 solana-svm-fuzz-harness-instr = { path = "svm-fuzz-harness/instr", version = "=3.1.0" }
+solana-svm-fuzz-harness = { path = "svm-fuzz-harness", version = "=3.1.0" }
 solana-svm-log-collector = { path = "svm-log-collector", version = "=3.1.0" }
 solana-svm-measure = { path = "svm-measure", version = "=3.1.0" }
 solana-svm-timings = { path = "svm-timings", version = "=3.1.0" }
