+++ svm-fuzz-harness/src/fixture/account_state.rs
