+++ svm-fuzz-harness/src/fixture/error.rs
