+++ svm-fuzz-harness/src/fixture/instr_context.rs
