+++ svm-fuzz-harness/src/instr.rs
@@ -5,12 +5,8 @@
 
 #![allow(clippy::missing_safety_doc)]
 
-pub mod program_cache;
-pub mod sysvar_cache;
-
 use {
     agave_precompiles::{get_precompile, is_precompile},
-    prost::Message,
     solana_account::AccountSharedData,
     solana_compute_budget::compute_budget::{ComputeBudget, SVMTransactionExecutionCost},
     solana_hash::Hash,
@@ -36,7 +32,7 @@ use {
         transaction_accounts::TransactionAccount, IndexOfAccount, InstructionAccount,
         TransactionContext,
     },
-    std::{collections::HashSet, ffi::c_int},
+    std::collections::HashSet,
 };
 
 /// Implement the callback trait so that the SVM API can be used to load
@@ -322,31 +318,6 @@ pub fn execute_instr_proto(input: ProtoInstrContext) -> Option<ProtoInstrEffects
     };
     let instr_effects = execute_instr(instr_context);
     instr_effects.map(Into::into)
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn sol_compat_instr_execute_v1(
-    out_ptr: *mut u8,
-    out_psz: *mut u64,
-    in_ptr: *mut u8,
-    in_sz: u64,
-) -> c_int {
-    let in_slice = std::slice::from_raw_parts(in_ptr, in_sz as usize);
-    let Ok(instr_context) = ProtoInstrContext::decode(in_slice) else {
-        return 0;
-    };
-    let Some(instr_effects) = execute_instr_proto(instr_context) else {
-        return 0;
-    };
-    let out_slice = std::slice::from_raw_parts_mut(out_ptr, (*out_psz) as usize);
-    let out_vec = instr_effects.encode_to_vec();
-    if out_vec.len() > out_slice.len() {
-        return 0;
-    }
-    out_slice[..out_vec.len()].copy_from_slice(&out_vec);
-    *out_psz = out_vec.len() as u64;
-
-    1
 }
 
 #[cfg(test)]
